+++ h/app.py
@@ -129,7 +129,6 @@ def includeme(config):
     # Site modules
     config.include("h.accounts")
     config.include("h.links")
-    config.include("h.nipsa")
     config.include("h.notification")
 
     # Configure sentry
